Serve the qcow images via apache
PCCI runs tests using qcow images that trust the vagrant insecure key.
These images are reasonably hard to rebuild for users, so we can serve
the images over https in case a user wants to use them in local
development. This also allows the users to inspect them for packages etc
etc.